ASL_ML_CLASSIFIER/requirements.txt
torch
torchvision
numpy
Pillow
tqdm
opencv-python
mediapipe
